feat: 건의/감사 글의 좋아요 수 계산 로직 추가

+++ src/main/java/EBus/EBusback/domain/post/dto/PostDetailResponseDto.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class PostDetailResponseDto {
 
 	private PostCreateResponseDto post;
-	private Long heartCount;
+	private Integer heartCount;
 	private PostMemberDto member;
 }
